Add remove_artists_tags operation
This removes ARTISTS and ALBUMARTISTS.

I prefer to include multiple instances of the ARTIST and ALBUMARTIST tags (e.g. I
split artists in fb2k), and so I am not managing these tags.

Kodi, and possibly other software, reads them, so it might be wreaking
havoc with how albums are split in my library.